+++ src/app/Router.tsx
@@ -8,7 +8,7 @@ import NotFound from '@/pages/notFound';
 import SigninPage from '@/pages/signin';
 import SignupPage from '@/pages/signup';
 import Layout from '@/widgets/layout/ui/MainLayout';
-import { ROUTES } from '@/shared/constants/routes';
+import { ROUTES } from '@/shared/model/constants/routes';
 
 interface AuthRouteProps {
   isPublic?: boolean;
