+++ src/features/auth/ui/OAuthCallBack.tsx
@@ -2,7 +2,7 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { useEffect } from 'react';
 import { useMutation } from '@tanstack/react-query';
-import { ROUTES } from '@/shared/constants/routes';
+import { ROUTES } from '@/shared/model/constants/routes';
 import { authQueries } from '../api/queries';
 
 export const OAuthCallback = () => {
