+++ src/features/auth/ui/WithdrawModal.tsx
@@ -5,8 +5,8 @@ import { useMutation } from '@tanstack/react-query';
 import { useAuthStore } from '@/shared/model/store/authStore';
 import { useModalStore } from '@/shared/model/store/modalStore';
 import { useSocketStore } from '@/shared/model/store/socketStore';
-import { MODAL } from '@/shared/constants/modal';
-import { ROUTES } from '@/shared/constants/routes';
+import { MODAL } from '@/shared/model/constants/modal';
+import { ROUTES } from '@/shared/model/constants/routes';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
 import { authQueries } from '../api/queries';
 import { type WithdrawFormData, withdrawSchema } from '../model/schema';
