+++ src/features/directMessage/ui/DMItem.tsx
@@ -3,8 +3,8 @@ import { toast } from 'react-toastify';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { DMUser } from '../model/types';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
-import { ROUTES } from '@/shared/constants/routes';
 import { cn } from '@/shared/lib/utils/utils';
+import { ROUTES } from '@/shared/model/constants/routes';
 import { UserAvatar } from '@/shared/ui/UserAvatar';
 import { DMQueries } from '../api/queries';
