+++ src/features/directMessage/ui/UnreadMessageIndicator.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from 'react-router-dom';
 import { useUnreadMessagesStore } from '@/shared/model/store/unreadMessagesStore';
-import { ROUTES } from '@/shared/constants/routes';
 import { DiscordIcon } from '@/shared/icons/DiscordIcon';
 import { cn } from '@/shared/lib/utils/utils';
+import { ROUTES } from '@/shared/model/constants/routes';
 
 export const UnreadMessageIndicator = () => {
   const unreadUsers = useUnreadMessagesStore((state) => state.unreadUsers);
