+++ src/features/server/ui/ChannelItem.tsx
@@ -5,12 +5,12 @@ import type { Channel } from '../model/types';
 import { useModalStore } from '@/shared/model/store/modalStore';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
-import { MODAL } from '@/shared/constants/modal';
-import { ROUTES } from '@/shared/constants/routes';
 import { AddUserIcon } from '@/shared/icons/AddUserIcon';
 import { ChannelUnlockedIcon } from '@/shared/icons/ChannelUnlockedIcon';
 import { VoiceIcon } from '@/shared/icons/VoiceIcon';
 import { WasteBasketIcon } from '@/shared/icons/WasteBasketIcon';
+import { MODAL } from '@/shared/model/constants/modal';
+import { ROUTES } from '@/shared/model/constants/routes';
 import { serverQueries } from '../api/queries';
 import { ChannelAddBtn } from './ChannelAddBtn';
