+++ src/features/server/ui/InvitedMemberModal.tsx
@@ -6,9 +6,9 @@ import { useInfiniteQuery, useMutation, useQuery } from '@tanstack/react-query';
 import { useModalStore } from '@/shared/model/store/modalStore';
 import { queryClient } from '@/shared/api/queryClient';
 import { QUERY_KEYS } from '@/shared/api/queryKeys';
-import { MODAL } from '@/shared/constants/modal';
 import { useDebounce } from '@/shared/lib/hooks/useDebounce';
 import { useInfiniteScroll } from '@/shared/lib/hooks/useInfiniteScroll';
+import { MODAL } from '@/shared/model/constants/modal';
 import { SearchInput } from '@/shared/ui/SearchInput';
 import { ModalContainer } from '@/shared/ui/layout/ModalContainer';
 import { friendQueries } from '../../friend/api/queries';
