+++ src/features/server/ui/ServerDropDownMenu.tsx
@@ -1,8 +1,8 @@
 import { useModalStore } from '@/shared/model/store/modalStore';
-import { MODAL } from '@/shared/constants/modal';
 import { AddUserIcon } from '@/shared/icons/AddUserIcon';
 import { ChannelUnlockedIcon } from '@/shared/icons/ChannelUnlockedIcon';
 import { SettingIcon } from '@/shared/icons/SettingIcon';
+import { MODAL } from '@/shared/model/constants/modal';
 import {
   DropdownMenuGroup,
   DropdownMenuItem,
