+++ src/features/user/ui/UserProfileCard.tsx
@@ -1,7 +1,7 @@
 import { useAuthStore } from '@/shared/model/store/authStore';
 import { useModalStore } from '@/shared/model/store/modalStore';
-import { MODAL } from '@/shared/constants/modal';
 import { EditIcon } from '@/shared/icons/EditIcon';
+import { MODAL } from '@/shared/model/constants/modal';
 import { UserAvatar } from '@/shared/ui/UserAvatar';
 import { SignOutButton } from '../../auth/ui/SignOutButton';
 import { WithdrawButton } from '../../auth/ui/WithdrawButton';
