+++ src/pages/channels/index.tsx
@@ -2,8 +2,8 @@ import { useModalStore } from '@/shared/model/store/modalStore';
 import { ChannelList } from '@/widgets/channel/ui/ChannelList';
 import { EditServerModal } from '@/features/server/ui/EditServerModal';
 import { InvitedMemberModal } from '@/features/server/ui/InvitedMemberModal';
-import { MODAL } from '@/shared/constants/modal';
-import { SEO_CONFIG } from '@/shared/constants/seo';
+import { MODAL } from '@/shared/model/constants/modal';
+import { SEO_CONFIG } from '@/shared/model/constants/seo';
 import { Seo } from '@/shared/ui/Seo';
 
 const ChannelPage = () => {
