+++ src/shared/model/constants/modal.ts
