+++ src/shared/model/constants/routes.ts
