+++ src/shared/model/constants/seo.ts
