+++ src/shared/ui/ChannelMessageDefault.tsx
@@ -1,8 +1,8 @@
 import { PropsWithChildren } from 'react';
 import { useModalStore } from '../model/store/modalStore';
-import { MODAL } from '../constants/modal';
 import { ArrowRight } from '../icons/ArrowRight';
 import { NewMemberIcon } from '../icons/NewMemberIcon';
+import { MODAL } from '../model/constants/modal';
 
 interface ChannelMessageDefaultProps extends PropsWithChildren {
   serverName: string | undefined;
