+++ src/widgets/sideNav/SideNav.tsx
@@ -5,8 +5,8 @@ import { serverQueries } from '@/features/server/api/queries';
 import { CreateServerAvatar } from '@/features/server/ui/CreateServerAvatar';
 import { JoinServerAvatar } from '@/features/server/ui/JoinServerAvatar';
 import { JoinToServerModal } from '@/features/server/ui/JoinToServerModal';
-import { MODAL } from '@/shared/constants/modal';
-import { ROUTES } from '@/shared/constants/routes';
+import { MODAL } from '@/shared/model/constants/modal';
+import { ROUTES } from '@/shared/model/constants/routes';
 import { DMAvatar } from '../../features/directMessage/ui/DMAvatar';
 import { ServerAvatar } from '../../features/server/ui/ServerAvatar';
 import { CreateServerModal } from './CreateServerModal';
